remove _didv2.py, rename remaining files
