+++ qetpy/core/didv/__init__.py
@@ -1,4 +1,3 @@
 from ._base_didv import *
 from ._didv import *
-from ._didv3 import *
-from ._didv2 import *
+from ._didv_priors import *
